:dizzy:[update]cancel_game()をcloseModal()から呼ぶ
伴い、module構造調整
