+++ frontend/static_vol/js/modules/modal.js
@@ -1,5 +1,8 @@
 'use strict';
 
+import { cancel_game } from "./match.js";
+import { initToken } from "./token.js";
+
 const endIndicator = (ev) => {
     const indicatorBar = ev.target;
     indicatorBar.removeEventListener('transitionend', endIndicator);
@@ -14,10 +17,8 @@ const showModal = (elHtml) => {
     elModal.innerHTML = elHtml;
 
     //キャンセルボタンにaddEventListener
-    const btnCancel = document.querySelectorAll('.blockBtnCancel_button');
-    btnCancel.forEach((btn) => {
-        btn.addEventListener('click', closeModal);
-    });
+    const btnCancel = document.querySelector('.blockBtnCancel_button');
+    btnCancel.addEventListener('click', closeModal);
 
     //インディケータがあれば進行、終了でcloseModal
     const indicator = document.getElementById('indicator');
@@ -33,20 +34,26 @@ const showModal = (elHtml) => {
 }
 
 const closeModal = () => {
-    //キャンセルボタンremoveEventListener
-    const btnCancel = document.querySelectorAll('.blockBtnCancel_button');
-    btnCancel.forEach((btn) => {
-        btn.removeEventListener('click', closeModal);
-    });
-    const indicator = document.getElementById('indicator');
-    if (indicator) {
-        const indicatorBar = indicator.querySelector('.unitIndicator_bar');
-        indicatorBar.removeEventListener('transitionend', endIndicator);
-    }
-    //modal close
-    const elModal = document.getElementById('wrapModal');
-    elModal.classList.remove('is-show');
-    elModal.innerHTML = '';
+    initToken()
+        .then((accessToken) => {
+            //キャンセルボタンremoveEventListener
+            const btnCancel = document.querySelector('.blockBtnCancel_button');
+            btnCancel.removeEventListener('click', closeModal);
+            //indicator removeEventListener
+            const indicator = document.getElementById('indicator');
+            if (indicator) {
+                const indicatorBar = indicator.querySelector('.unitIndicator_bar');
+                indicatorBar.removeEventListener('transitionend', endIndicator);
+            }
+            //cancel game
+            cancel_game(accessToken);
+        })
+        .then(() => {
+            //modal close
+            const elModal = document.getElementById('wrapModal');
+            elModal.classList.remove('is-show');
+            elModal.innerHTML = '';
+        });
 }
 
 export { showModal, closeModal };
